requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.0.3
pyarrow==12.0.1
pyyaml==6.0.1
python-multipart==0.0.6

# Required for model loading
joblib==1.3.2

# Data collection
ccxt==4.3.89
python-dotenv==1.0.0

# ML dependencies
xgboost==2.0.2
scikit-learn==1.3.2
# Optional ML dependencies (uncomment as needed)
#tensorflow==2.15.0

# Deep Learning dependencies
torch==2.1.0

